Fix warn_unused_result warning.
A warning triggered by attribute warn_unused_result cannot be suppressed by
casting the result to void. Properly check the return value and pass it to the
caller.

Fix the check for the number of bytes written. The return value of fread is the
number of items written, not the number of bytes.

Change-Id: I8c60e23ac549e22fd3f7dd5c51c5a8e6fc517cb1