Update patch set 2
Patch Set 2:

I havnt checked binsize, but why would it be any bigger? pointer dereferences are worse than array indexing?

Anyway, this isnt about dynamic lcd sizing and I'm not doing any work in this change for it. this is purely to lay down groundwork to make that (and other future ideas) easier.

Patch-set: 2