Revise the PCM callback system after adding multichannel audio.
Additional status callback is added to pcm_play/rec_data instead of
using a special function to set it. Status includes DMA error
reporting to the status callback. Playback and recording callback
become more alike except playback uses "const void **addr" (because
the data should not be altered) and recording  uses "void **addr".
"const" is put in place throughout where appropriate.

Most changes are fairly trivial. One that should be checked in particular
because it isn't so much is telechips, if anyone cares to bother. PP5002 is
not either but that tested as working.

Change-Id: I4928d69b3b3be7fb93e259f81635232df9bd1df2