Update patch set 1
Patch Set 1:

the fuze+ touchpad seems to be fairly different from one device to another and/or the way people use it change a lot its way to react. This is the reason why a single two value setting like on the gigabeats is not really meaningful here.
I couldn't test the patch myself for the moment (we do it soon): my device is dead and I had to order a new one.
This will be splitted into general rockbox change and fuze+ part if the principle is accepted - for the moment just testing - comment about implementation, general idea, and code are welcome.

Patch-set: 1
Label: Code-Review=0
Label: Tested=0