Added directory functions to Lua API
Implemented here are:
opendir(path) -- returns 1 or nil on failure
readdir()     -- returns a string or nil on failure/end of dir
closedir()    -- returns 1 or nil on failure

The method is a little hackish; you can only have one directory open at
a time, and the directory pointer isn't exposed to Lua at all. If
someone wants to make it act more like the C version later, just keep my
method as a default if no directory pointer is specified -- that way
code written for this version will continue to work.

Change-Id: Ibeee39bdf44bcbc64c65252703ea1f1b1321faf6