Update patch set 2
Patch Set 2:

I think the hidden flag toggling would be better placed in the properties plugin -- the plugin displays the values, so it makes sense to toggle it at that place as well. Furthermore, I don't consider changing the flag a core functionality which is also a reason to put it in a plugin.

Patch-set: 2
Label: Code-Review=0
Label: Tested=0