List Bookmarks: Set %cs (Current Screen) to "Bookmark browser" when listing bookmarks from the Context Menu
Currently when you select list bookmarks from the context menu %cs returns "Context Menu" when in the bookmark browser screen.  This change makes %cs return "Bookmark browser" when listing bookmarks from the context menu, the same as when you list recent bookmarks.  This change will make it possible to determin that you are on a bookmark browser screen when skinning using sbs files.

Change-Id: I7fb93525fbafb5d14bba2ae5df7a78df908d09ae