Update patch set 1
Patch Set 1:

You understand that the "setting" is *never* read right? it is only there because the macros in settings_list.c can't handle NULL for the setting pointer.

In fact, I'd argue that this change is dangerous in that it makes people think that changing the value will actually affect anything - which it doesnt.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0