Update patch set 1
Patch Set 1:

I understand that this is a "synthetic" setting, i.e. its value is not written to the config file directly. But it is still a variable in the program, and hence should be self explaining as much as possible (type and name).

Patch-set: 1
Reviewer: Gerrit User 1000045 <1000045@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0