Play music or complex tones with square wave channel.
Perhaps the stupidest thing in the world so far. By defining an
array a frequencies, durations and amplitudes that can be passed
to beep_play_melody, it is possible to play music or perhaps
indicators other than simple raw beep.

At this time, all splashes play Dixie a la The General Lee just
as a demo. Of course that isn't meant to be committed. ;-)

Change-Id: I1e89b3f3682d9c03e8fdbcd91e3deaa1558a9ae6