Constrain get_next_dir() to directories below global_settings.start_directory.
If the user has set "Start File Browser Here" (config.cfg: start directory)
to anything other than root and "Auto-Change Directory" is set to "Yes" or
"Random", the directory returned when an auto change is required will be
constrained to the value of "start directory" or below.

Edit: Actually include the stripped trailing whitespace this time.

Change-Id: Iaab773868c4cab5a54f6ae67bdb22e84642a9e4b