Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Thanks for your work. I'm sorry, but the way this is done is completely pointless and defeats the purpose of notifying users about changes. I guess I was too unspecific when mentioning a changlog dialog :(

This implementation puts a changelog in yet another hidden place users are unlikely to see. What I'd like to see is users getting notified of the changes when they start Rockbox Utility (or start a new version for the first time). Right now we have some "this is a new installation or new version" dialog. A changelog should replace this, so users immediately get shown what's new -- perhaps some important bug has been fixed they should get notified about (I'm thinking about the constant claiming that TTS on Windows is broken, which definitely isn't the case anymore).

For this a new widget is necessary (i.e. something like "class ChangeLog : public QWidget"). This widget should hold
- a title line
- a QTextBrowser for the formatted text
- a "don't show this dialog again" checkbox. Maybe rather a "show this info on each start".
- a Close button.

The changelog widget should provide functions for setting the text to display, telling it if to interpret the text as some kind of markup or not, send a signal on close so the caller can destroy the object and some means to handle the checkbox. Also, I'd like to see some kind of simple markup that allows feeding it a plain text document that gets formatted by the widget (though supporting markup should probably done separately since it's more work and requires to agree on how the markup is supposed to look like. I'm thinking of something similar to markdown or restructured text, trimmed down to our needs and with some extensions like replacing special OS type markers with nice images. The idea is to move the changelog out of the wiki afterwards and just link to the file in gitweb instead, so the format should be easy readable in plain text. It also would be good to have unit tests for the markup parser).

As for the patch itself, you cannot simply read a file -- this doesn't work when deploying Rockbox Utility as single binary, since we don't want to ship additional files. Instead you need to put the changelog text file into the resources and read it from there.

Patch-set: 1