Display function names in backtrace using -mpoke-function-name
Use -mpoke-function-name to have the compiler put the function name
just before the prologue with some magic value. To implement this
properly we need to tell if can read from a memory location (this
is also useful for the backtracer). Implement this by removing
all implementations of exception handler from the various crt0.S
and put one reference on in system-arm.c which is weak and
overridden by the unwinder when compiled in to intercept data
abort and not die when trying to poke at some memory location.

Change-Id: I3e1d5225a4f03df2b8245925a303a7d3a948db1a