Update patch set 1
Patch Set 1:

Isn't this a waste of memory?
Different ways of debugging panics could be a) put the symbol table next to the Rockbox binary on-disk and load it on-demand upon a panic and/or b) dump the stack + relevant memory values to disk or even c) do a full memory dump.

Patch-set: 1
Reviewer: Gerrit User 1000010 <1000010@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0