Update patch set 1
Patch Set 1:

I'm willing to pay this amount of memory if it speeds up debugging. This is very useful for developers. I agree that it should probably be made optional (like Logf and others options).
Your propositions have a number of drawbacks: if the panic comes from the storage system, you can't read/write to the disk. If you load the symbols from the disk, you use the same amount of memory. This solution is simple to implement and let the compiler to most of the job. You are free to implement other mechanisms so we can even choose between them at configuration time.

Patch-set: 1