Fundamentally rewrite some aspects of the audio DSP.
Creates a standard buffer passing system for processing stages.
Removes restrictions on the number of samples that can be processed at
a time and it pays attention to destination buffer size restrictions
without having to limit input count, which also allows pcmbuf to
remain fuller and safely set its own buffer limits as it sees fit.
Works in harmony with stages that are not deterministic in terms of
sample input/output ratio (like both resamplers but most notably
the timestretch).

Additionally fixes quirks with TDSpeed hanging up and it now operates
properly throughout its full settings range.

I still need to think about a few things because samples can linger
in the pipline. Maybe fix some things that look ugly to me if there's
a better way to do it. Definitely not quite ready for prime-time.

ASM resampling is disabled for the moment since it must be tweaked
back into compatibility. C routines were updated and should function
properly.

test_codec is busted because I haven't brought it into line yet.

Will split up miscellaneous changes into separate commits later as I
see fit to do. But, it's a broad addressing of issues with a main
focus and I don't personally care about "relevence". My workflow is
a flow and 'yall should be used to that by now. :-)

Change-Id: Ib206ec78f6f6c79259c5af9009fe021d68be9734