Fundamentally rewrite some aspects of the audio DSP.
Creates a standard buffer passing, local data passing and messaging
system for processing stages. Stages can be moved to their own source
files to reduce clutter and ease assimilation of new ones.

Formats and change notifications are passed along with the buffer so
that they arrive at the correct time at each stage in the chain
regardless of the internal delays of a particular one.

Removes restrictions on the number of samples that can be processed at
a time and it pays attention to destination buffer size restrictions
without having to limit input count, which also allows pcmbuf to
remain fuller and safely set its own buffer limits as it sees fit.

Works in harmony with stages that are not deterministic in terms of
sample input/output ratio (like both resamplers but most notably
the timestretch). As a result it fixes quirks with TDSpeed hanging
up and it now operates properly throughout its full settings range.

Maybe fix some things that look ugly to me if there happens to be a
better way to do it but I would consider this one a commit candidate.

Binsize increase isn't exactly modest but emphasis was placed upon
minimizing commonly run code in the main loops and saving IRAM, even
if it meant more support code. I will see if I can squeeze down
some sections. The large guardbuf for PCM disappears so really it's
a net decrease in resource usage.

Quick tests on coldfire with SPC showed a -1.24 MHz (+1.47%)
performance benefit (using only the resampler) over HEAD and
-1.20 MHz (+1.43%) over this resampler implementation in an older
patch which isn't exactly insignificant.

Will split up miscellaneous changes into separate commits later as I
see fit to do. But, it's a broad addressing of issues with a main
focus and I don't personally care about "relevence" at the moment.
My workflow is a flow and y'all should be used to that by now. :-)

Change-Id: Ib206ec78f6f6c79259c5af9009fe021d68be9734