Fundamentally rewrite much of the audio DSP.
Creates a standard buffer passing, local data passing and messaging
system for processing stages. Stages can be moved to their own source
files to reduce clutter and ease assimilation of new ones. dsp.c
becomes dsp_core.c which supports an engine and framework for effects.

Formats and change notifications are passed along with the buffer so
that they arrive at the correct time at each stage in the chain
regardless of the internal delays of a particular one.

Removes restrictions on the number of samples that can be processed at
a time and it pays attention to destination buffer size restrictions
without having to limit input count, which also allows pcmbuf to
remain fuller and safely set its own buffer limits as it sees fit.
There is no longer a need to query input/output counts given a certain
number of input samples; just give it the sizes of the source and
destination buffers.

Works in harmony with stages that are not deterministic in terms of
sample input/output ratio (like both resamplers but most notably
the timestretch). As a result it fixes quirks with timestretch hanging
up with certain settings and it now operates properly throughout its
full settings range.

Binsize increase isn't exactly modest but emphasis was placed upon
minimizing commonly run code in the main loops and saving IRAM, even
if it meant more support code. I will see if I can squeeze down
some sections. The large guardbuf for PCM disappears so really it's
a net decrease in resource usage. Thumb code, on test builds, only
increases a very small amount, much less than expected (a factor
of 10 difference - why?).
Change-Id: Ib206ec78f6f6c79259c5af9009fe021d68be9734