This patch adds radio support to Samsung YP-R0
It doesn't include the kernel module source code I needed to write in order to have direct access to I2C bus, but it will. This is the first test commit.
Basically it uses the default SI4700 radio chip driver, the only thing that's different is the I2C access, written specifically to interact with my kernel module.
The overhead is reduced since there is just one sys call (ioctl) when writing (reading) one or more registers at once.
Next things to add are:
- fix the radio context keymappings
- add to the build system my kernel module, which should be built when building the modified firmware to launch rockbox
- RDS support!
- Cleanups, in case ;)

Change-Id: I0ed125641e00f93124d7a34f90dd508e7f1db5a4
Signed-off-by: Lorenzo Miori <memorys60@gmail.com>