Check for existing firmware file on H100 / H300.
Show an error if copying the firmware file to the player failed. Also, check if
there is already a firmware file present on the player. If so abort and tell
the user about it, since QFile::copy() doesn't overwrite existing files and the
user might not expect the file to get overwritten.

Change-Id: Ie63963289965900a4ab519ddf128246f89d81c6c