Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; Fails on target

This isn't going to fix anything...in fact, this will break things horribly.
I brought this up on IRC, but I forgot to comment here and I just saw that the task was still up...so, I'm doing so.
You are most certainly not supposed to be able to flash an unencrypted bootloader with ipodpatcher.
Its a bit of a chicken/egg scenario where the bootloader must first be encrypted with crypt_firmware on a functional (as in, boots Rockbox) Nano2G before flashing.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=-1
Label: Tested=-1