Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I can't test this on target, but this looks like an excellent patch. You can back it up with a reference to the specification and it actually fixes a problem in the field.

What's next, a few more tests with this code and other kind of drives to verify it doesn't break anything?

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=+1
Label: Tested=0