Update patch set 1
Patch Set 1:

I've only implemented this for the sim and target so far, app_open() needs handling also.

Also, the example useable in playlist.c isnt really finished. It will still fail if the file is further in the playlist and not yet fully buffered if it isnt buffered at all it wil be skipped without issue)

Patch-set: 1
Label: Code-Review=0
Label: Tested=0