Update patch set 3
Patch Set 3:

Is playback crashing if the file is lost? It should simply truncate. If that's not working right it should probably be fixed.

Commenting on your source comment: Because it's not the last track doesn't mean buffering is finished for it. That "last track" is just the one where all the handles haven't been created or it's still empty. Handles simply allocate space. See audio_on_handle_finished where it notes that it doesn't even know what's coming next. I'd recommend implement these things in buffering which does deal with files directly and provide a buffering event to playback in case of unrecoverable file errors/changes.

Patch-set: 3
Label: Code-Review=0
Label: Tested=0