Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

This is undoing to some degree the compartmentalizing of processing implementations. It's sort of on the line but there's alot of pressure to rearrange things in this area and it should work afaict so I'll +1.

Patch-set: 3
Reviewer: Gerrit User 1000041 <1000041@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=+1
Label: Tested=0