Introduce new hermite polynomial resampler.
Currently, the linear resampler is left in as a compile time option.
Eventually we may want to remove it entirely depending on how fast the
hermite resampler is on older targets as the quality is much better using
hermite.

Based on work in g#231 by Michael Sevakis and I, but remove real networks
code in favor of my own implementation.

Added a temporary setting to set the resampler type (for quick comparison)
and renamed everything to "resample..."  instead of "lin_resample...".

The latter will likely be done separately and I doubt the former would get
pushed.

Change-Id: Ic45d84bc66c5b312ef373198297a952167a4be26