Update patch set 6
Patch Set 6:

Well, he's no Jan Brady. Anyway, I wasn't sure about the sentiments of actually committing this to the core so it "look[ed] good to me". If it is wanted, then I'll happily approve and submit it. Looking at the equations, it seems it would be a bit tricky to make fully configurable and replace the existing crossfeed. I wonder if an all-pass filter could compensate the combing effect in the current one. One great thing about this one is the cheapness of computation.

Patch-set: 6