Update patch set 1
Patch Set 1:

I'm not completely decided about this.

The RDS spec says that all segments should arrive in order. This patch ignores that in favour of an arbitrary timeout.

I think this patch makes it both more likely to receive an PS/RT message but also more likely that it is not actually a valid message (for example mixed data from two different messages).
On the other hand you could argue that a PS message is sent quite often so a temporary invalid PS is not such a big deal, and an RT message is already "protected" by the A/B flag.

Are you sure this is not merely hiding some underlying problem, like reading out the RDS data too slowly or too late (e.g. on the wrong edge of the interrupt)?

Patch-set: 1