Update patch set 1
Patch Set 1:

The specification is a bit vague, it also says everything should be sent twice to make it more likely to arrive... The callback is triggered every time the interrupt pin is 0. It's certainly a bit slow because it's handled by a thread but we can't do bit toggling in the interrupt handler. Anyway, that's basically the same code as the as3525.

Patch-set: 1
Reviewer: Gerrit User 1000040 <1000040@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0