Merge some changes from FS#10320 - Timestretch: Update algorithm
Not dedicated to the change. The Soundtouch params work pretty well
for modestly retempoing music but do not work so well for extreme
timestretching of voice (which will sound like two voices overlapped
onto one another if speeding up by say, 180%). The current algorithm
is defintely superior for that.

I'm not convinced the type of autocorrelation would have a huge
impact on quality so I did not adopt anything in that area. Nothing
really seems different if not using the Soundtouch params.

I used huge buffers to be sure for now and I would actually
calculate the needed space if pushing this.

Change-Id: I7e862306122eaa94a9185b557e5ce46e77bd851f