Add real waveform view to oscilloscope plugin. (FTL)
Adds some additional niftyness like a floating popup display that
could be pulled into the plugin library (which if/when this is pushed,
I'll definitely do). Speed changes are now gradual for both views.

Refined a few behavioral things overall.

It needs a bit of help from a direct PCM channel callback so it may
capture PCM for waveform display. Also need a few other core routines
to help out.

Messes with some keymaps. Some targets need keymaps. Some devices
can't support the addition view because it requires a large buffer
for samples and the plugin buffer is small but they can still use
the popup display since the plugin is much smaller in that case.

Change-Id: Ieb5b7922a2238264e9b19a58cb437739194eb036