Update patch set 2
Patch Set 2:

Firstly, I'm not playing around with this/that yield stuff anymore. I'm a bit perplexed of the perpetual aversion to proper thread code. Won't do it. It's pissing me off to no end. I don't want to track down bugs that arise out of some old bygone days of "simplicity". I'll pack up and leave if that's how we're going to continue. Second, It is accessed in a multithreaded manner, yielding within buffer callbacks. Nothing about buflib itself appears reentrant, which is fine if using it on one thread. Calls to alloc are possible during callbacks, among other things. Multithreaded use of core alloc in that way will increase soon since I've decided to go right on overboard with this and add buffer mangement for the audiobuffer. What we have is inadequate to remain robust against constantly changing code.

Patch-set: 2
Label: Code-Review=0