Update patch set 2
Patch Set 2:

I thought about putting it in buflib but then I realized that it would be pointless if you're using it in as a simple allocator on a single thread in a plugin or some other case. I wouldn't want all contexts sharing the same one either and the mutex must be in sharedbss even if the object is not used with threads of multiple cores because other cores can interact somewhat invisibly with them because of priority inheritance. If buflib was meant to be core allocator only, then why split it in the manner it is?

Patch-set: 2