Update patch set 2
Patch Set 2:

What I'm doing next as part of a main buffer overhaul is to manage it with simple ownership. No other may steal, move or shrink the buffer without permission of the owner and allowed ops are specified upon obtaining it. Whoever is using it is responsible for dealing with ops. Playback will be nothing but another user and have no special status.

Patch-set: 2