ypr0: Add kernel module to required support fm radio.
To support fm radio a kernel module was written. This module is added to
the patched firmware that loads Rockbox. It's pre-compiled but its source
archive provided.

The kernel module provides raw-access to the built-in si4709 radio chip.
Our existing si 47xx drivers can be re-used this way.

The module itself was written Lorenzo Miori, I only integrated it
into the tree.

Change-Id: I6205d28a505d57791eaeb627e6856b9a1eaeaeaa