Filter LaTeX output for errors.
Pipe the output of LaTeX through a Perl script and filter out information that
is irrelevant in deciding if building the manual actually worked. Format errors
in a similar way to gcc output to allow existing scripts catching it.

Enabling verbose output during the make run will not remove parts of the output
but only do some reflowing. The full log is always available in the manual
subfolder.

Change-Id: I15d35b4d3c73fafe2a4357168ca8ada51355f221