Update patch set 3
Patch Set 3:

OK, I've just tried the patch again and I have concerns... now there is no way to know if the timer is running. I also suspect that the "start sleep timer" 'setting' doesnt even work correctly (i.e it should restart if you enter it when it is already set to yes.

To fix that issue what should happen is on confirmation (i.e setting the setting to YES) the timer starts, there is some visual indication of it running, and then the setting gets reset to OFF. err.. except then this means to disable the setting you need to first reenable it and then disable it.

Perhaps the setting should really be (with a custom formatter) "Off", "Continue (*time remaining*)", "Start/Run (*time*)" instead. then when it is enabled you magically switch the setting to "continue".

I'd also like an item in the statusbar for the remaining time (and a skin tag if there isnt one already) but these could/should happen in a separate change.

Patch-set: 3
Reviewer: Gerrit User 1000011 <1000011@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0