Make playlist.c aware of external storage when determining paths
format_track_path currenyly corrects for things like windows volumes  and
relative paths in playlists, but does not handle external media like SD
cards correctly, resulting in some seemingly valid playlists not working
because of rockbox's mount point for external media.  Correct this by checking
to see if a playlist is on external media and then formulate the path correctly
if it is.

Unfortunately, this breaks the playlist_save logic if the CWD is on an external
device.  Its not clear to me why we should be checking the CWD when saving a
playlist, as the only apparent use of this is to let people save relative
paths on the virtual keyboard.  As far as I can tell, this is actually more
difficult to do then using an absolute path given that we insert the CWD
onto the virtualkeyboard by default.  Therefore, I'm removing the option
to use '..' in playlist save paths since its seems useless.

Change-Id: I47946cc45d776c7a72ecbd0ecc720dbf85550f6f
Reviewed-on: http://gerrit.rockbox.org/270
Reviewed-by: Michael Giacomelli <mgiacomelli@gmail.com>
Tested-by: Michael Giacomelli <mgiacomelli@gmail.com>