Make the SDL app stream to an icecast2 server.
Early work, not configurable yet (neither build not settings), so
not at all ready for master.

This updates pcm-gstreamer.c to be less maemo-specific, and adds a
gstreamer pipe to it to stream to icecast2 for the non-maemo case.
(some) metadata is also added to the stream, but a better way to do
that needs to be found.

Change-Id: I38a4d25dee7533f4edcb1a56a1585cb258675632