Drop size entries from voicefiles.
The voicefile format has a header that contains the offset and size of each
clip. Since clips are consecutive and have to be ordered there is no real need
for a separate size field in the header, it can be calculated as difference to
the following clip.

This reduces the size of a typical voicefile by about 2.8kiB (~700 voice
entries * 4 bytes per clip) in trade with an additional "end" entry that simply
points to an empty clip.

Update voicefile to support multiple versions of the voice file format and bump
version to 500.

Change-Id: Ied6d43e3417bb1d86467861a947dc870147dac54