sleeptimer now has TWO menuitems: 	Start Sleep Timer / Shutdown in xx min 	Off               / Cancel Sleep Timer Depending on the sleeptimers status.
sleeptimer_on_startup only used in settings_applied_on_boot

To restart the sleeptimer you have to cancel the current sleeptimer and start it again
which is easy enough in QS and Menu.

There is a problem in QS however: If Menuitemstring have to scroll then switching seems to
be broken.

I put F_TEMPVAR in settings_list.c but could not see any difference in behaviour.

Change-Id: Iea1e18ac319aaebe1682c8353b40918e5f5488a5