Update patch set 5
Patch Set 5: (1 inline comment)

I was testing the QuickScreen so I did not notice its not working in the Menu. And herein lies a problem:

There is a difference in QS and Menu. in QS settings are applied on exit.
In the Menu settings don´t get applied, and the sleeptimer function is not called, so nothing happens.

Adding a callback to the sleeptimer makes the menu work, but in QS it will instantly takeover new settings if you cycle through the 3 states. After OFF comes CONTINUE.. but this will always be 0 after OFF!

Do you have an idea to fix this problem?

I took the "Nothing to Resume" string because it was already defined. An alternative could add "Expired" to the lang entries.

Do you have a suggestion for settings_applied_on_startup? 
I don´t like it so much either. But how can the sleeptimer_on_boot setting be treated otherwise?

Patch-set: 5
Reviewer: Gerrit User 1000138 <1000138@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0